feat: 나머지 EntityClass 필요 로직 EntityPersister 이관 및 미사용 메소드 제거
