+++ src/main/java/hibernate/entity/EntityClass.java
@@ -6,11 +6,7 @@ import jakarta.persistence.Entity;
 import jakarta.persistence.Table;
 
 import java.lang.reflect.Constructor;
-import java.util.AbstractMap;
-import java.util.LinkedHashMap;
 import java.util.List;
-import java.util.Map;
-import java.util.stream.Collectors;
 
 public class EntityClass<T> {
 
@@ -57,31 +53,6 @@ public class EntityClass<T> {
             return clazz.getDeclaredConstructor();
         } catch (NoSuchMethodException e) {
             throw new IllegalStateException("기본 생성자가 존재하지 않습니다.");
-        }
-    }
-
-    public Map<EntityColumn, Object> getFieldValues(final Object entity) {
-        validateEntityType(entity);
-        return entityColumns.getValues()
-                .stream()
-                .map(entityColumn -> new AbstractMap.SimpleEntry<>(entityColumn, entityColumn.getFieldValue(entity)))
-                .filter(entry -> entry.getValue() != null)
-                .collect(Collectors.toMap(
-                        Map.Entry::getKey,
-                        Map.Entry::getValue,
-                        (existing, replacement) -> existing,
-                        LinkedHashMap::new
-                ));
-    }
-
-    public Object extractEntityId(final Object entity) {
-        validateEntityType(entity);
-        return getEntityId().getFieldValue(entity);
-    }
-
-    private void validateEntityType(final Object entity) {
-        if (clazz != entity.getClass()) {
-            throw new IllegalArgumentException("EntityClass와 일치하지 않는 객체입니다.");
         }
     }
