+++ src/main/java/hibernate/entity/EntityPersister.java
@@ -14,6 +14,7 @@ import java.util.stream.Collectors;
 
 public class EntityPersister<T> {
 
+    private final Class<T> clazz;
     private final EntityTableName tableName;
     private final EntityColumns entityColumns;
     private final JdbcTemplate jdbcTemplate;
@@ -22,12 +23,14 @@ public class EntityPersister<T> {
     private final UpdateQueryBuilder updateQueryBuilder = UpdateQueryBuilder.INSTANCE;
 
     public EntityPersister(final Class<T> clazz, final JdbcTemplate jdbcTemplate) {
+        this.clazz = clazz;
         this.tableName = new EntityTableName(clazz);
         this.entityColumns = new EntityColumns(clazz.getDeclaredFields());
         this.jdbcTemplate = jdbcTemplate;
     }
 
     public boolean update(final Object entity) {
+        validateEntityType(entity);
         EntityColumn entityId = entityColumns.getEntityId();
         final String query = updateQueryBuilder.generateQuery(
                 tableName.getTableName(),
@@ -39,6 +42,7 @@ public class EntityPersister<T> {
     }
 
     public void insert(final Object entity) {
+        validateEntityType(entity);
         final String query = insertQueryBuilder.generateQuery(
                 tableName.getTableName(),
                 getFieldValues(entity)
@@ -47,6 +51,7 @@ public class EntityPersister<T> {
     }
 
     public void delete(final Object entity) {
+        validateEntityType(entity);
         EntityColumn entityId = entityColumns.getEntityId();
         final String query = deleteQueryBuilder.generateQuery(
                 tableName.getTableName(),
@@ -67,5 +72,11 @@ public class EntityPersister<T> {
                         (existing, replacement) -> existing,
                         LinkedHashMap::new
                 ));
+    }
+
+    private void validateEntityType(final Object entity) {
+        if (clazz != entity.getClass()) {
+            throw new IllegalArgumentException("EntityClass와 일치하지 않는 객체입니다.");
+        }
     }
 }
