+++ src/test/java/hibernate/entity/EntityPersisterTest.java
@@ -15,6 +15,7 @@ import java.sql.ResultSet;
 import java.sql.SQLException;
 
 import static org.assertj.core.api.Assertions.assertThat;
+import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 class EntityPersisterTest {
 
@@ -56,6 +57,13 @@ class EntityPersisterTest {
     }
 
     @Test
+    void update_쿼리_실행_시_다른_타입의_entity가_입력될_경우_예외가_발생한다() {
+        assertThatThrownBy(() -> entityPersister.update(new ErrorEntity()))
+                .isInstanceOf(IllegalArgumentException.class)
+                .hasMessage("EntityClass와 일치하지 않는 객체입니다.");
+    }
+
+    @Test
     void insert_쿼리를_실행한다() {
         // given
         TestEntity givenEntity = new TestEntity("최진영");
@@ -66,6 +74,13 @@ class EntityPersisterTest {
 
         // then
         assertThat(actual).isEqualTo(1);
+    }
+
+    @Test
+    void insert_쿼리_실행_시_다른_타입의_entity가_입력될_경우_예외가_발생한다() {
+        assertThatThrownBy(() -> entityPersister.insert(new ErrorEntity()))
+                .isInstanceOf(IllegalArgumentException.class)
+                .hasMessage("EntityClass와 일치하지 않는 객체입니다.");
     }
 
     @Test
@@ -80,6 +95,13 @@ class EntityPersisterTest {
 
         // then
         assertThat(actual).isEqualTo(0);
+    }
+
+    @Test
+    void delete_쿼리_실행_시_다른_타입의_entity가_입력될_경우_예외가_발생한다() {
+        assertThatThrownBy(() -> entityPersister.delete(new ErrorEntity()))
+                .isInstanceOf(IllegalArgumentException.class)
+                .hasMessage("EntityClass와 일치하지 않는 객체입니다.");
     }
 
     private Integer testEntityCount() {
@@ -120,5 +142,8 @@ class EntityPersisterTest {
         public TestEntity(String name) {
             this.name = name;
         }
+    }
+
+    static class ErrorEntity {
     }
 }
